Bump version to 1.0.0-dev - Add Font Awesome - Update StaggeredGridView - Make AboutScreen Responsive
