+++ lib/main.dart
@@ -1,21 +1,16 @@
 import 'dart:io';
 
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:flutube/controller/internet_connectivity.dart';
 import 'package:bot_toast/bot_toast.dart';
 import 'package:hive_flutter/hive_flutter.dart';
-import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:lucide_icons/lucide_icons.dart';
 import 'package:path_provider/path_provider.dart';
-import 'package:salomon_bottom_bar/salomon_bottom_bar.dart';
 
+import 'package:flutube/homepage.dart';
 import 'package:flutube/utils/utils.dart';
 import 'package:flutube/models/models.dart';
-import 'package:flutube/screens/screens.dart';
-import 'package:flutube/widgets/widgets.dart';
 import 'package:flutube/providers/providers.dart';
+import 'package:flutube/controller/internet_connectivity.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -58,190 +53,6 @@ class MyApp extends HookConsumerWidget {
       themeMode: ref.watch(themeTypeProvider),
       debugShowCheckedModeBanner: false,
       home: MyHomePage(),
-    );
-  }
-}
-
-class MyHomePage extends HookConsumerWidget {
-  MyHomePage({Key? key}) : super(key: key);
-
-  final PageController _controller = PageController();
-
-  @override
-  Widget build(context, ref) {
-    final _currentIndex = useState<int>(0);
-    final extendedRail = useState<bool>(false);
-    final _addDownloadController = TextEditingController();
-
-    final mainScreens = [
-      const HomeScreen(),
-      const LikedScreen(),
-      const PlaylistScreen(),
-      const DownloadsScreen(),
-      const SettingsScreen(),
-    ];
-
-    final Map<String, List<IconData>> navItems = {
-      "Home": [LucideIcons.home],
-      "Liked": [LucideIcons.thumbsUp],
-      "Playlist": [LucideIcons.listPlus],
-      "Downloads": [LucideIcons.download],
-      "Settings": [LucideIcons.settings],
-    };
-
-    return Scaffold(
-      appBar: AppBar(
-        elevation: 0,
-        title: Row(
-          children: [
-            if (!context.isMobile) ...[
-              GestureDetector(
-                onTap: () => extendedRail.value = !extendedRail.value,
-                child: const Padding(
-                  padding: EdgeInsets.all(10.0),
-                  child: Icon(LucideIcons.menu),
-                ),
-              ),
-              const SizedBox(width: 15),
-            ],
-            Text(myApp.name),
-          ],
-        ),
-        actions: [
-          buildSearchButton(context),
-          if (_currentIndex.value == 3)
-            IconButton(
-              icon: const Icon(LucideIcons.trash2),
-              onPressed: () {
-                final deleteFromStorage = ValueNotifier<bool>(false);
-                showPopoverWB(
-                  context: context,
-                  builder: (ctx) => ValueListenableBuilder<bool>(
-                      valueListenable: deleteFromStorage,
-                      builder: (_, value, ___) {
-                        return Column(
-                          crossAxisAlignment: CrossAxisAlignment.start,
-                          children: [
-                            Text('Clear all items from download list?', style: context.textTheme.bodyText1),
-                            CheckboxListTile(
-                              value: value,
-                              onChanged: (val) => deleteFromStorage.value = val!,
-                              title: const Text("Also delete them from storage"),
-                            ),
-                          ],
-                        );
-                      }),
-                  onConfirm: () {
-                    final downloadListUtils = ref.read(downloadListProvider);
-                    for (DownloadItem item in downloadListUtils.downloadList) {
-                      if (File(item.queryVideo.path + item.queryVideo.name).existsSync() && deleteFromStorage.value) {
-                        File(item.queryVideo.path + item.queryVideo.name).deleteSync();
-                      }
-                    }
-                    downloadListUtils.clearAll();
-                    context.back();
-                  },
-                  confirmText: "Yes",
-                  title: "Confirm!",
-                );
-              },
-              tooltip: "Clear all",
-            ),
-          if (_currentIndex.value == 4)
-            PopupMenuButton(
-              itemBuilder: (context) {
-                return [
-                  PopupMenuItem(
-                    child: const Text('Reset default'),
-                    onTap: () => resetDefaults(ref),
-                  )
-                ];
-              },
-            ),
-          const SizedBox(width: 10),
-        ],
-      ),
-      body: Row(
-        children: [
-          if (!context.isMobile)
-            NavigationRail(
-              destinations: [
-                for (var item in navItems.entries)
-                  NavigationRailDestination(
-                    label: Text(item.key, style: context.textTheme.bodyText1),
-                    icon: Icon(item.value[0]),
-                    selectedIcon: Icon(item.value.length == 2 ? item.value[1] : item.value[0]),
-                  ),
-              ],
-              minExtendedWidth: 200,
-              extended: extendedRail.value,
-              selectedIndex: _currentIndex.value,
-              onDestinationSelected: (index) => _controller.jumpToPage(index),
-            ),
-          Flexible(
-            child: FtBody(
-              child: PageView.builder(
-                controller: _controller,
-                itemCount: mainScreens.length,
-                itemBuilder: (context, index) => mainScreens[index],
-                onPageChanged: (index) => _currentIndex.value = index,
-              ),
-            ),
-          ),
-        ],
-      ),
-      floatingActionButton: _currentIndex.value == 0
-          ? FloatingActionButton(
-              onPressed: () async {
-                if (_addDownloadController.text.isEmpty) {
-                  var clipboard = await Clipboard.getData(Clipboard.kTextPlain);
-                  var youtubeRegEx = RegExp(
-                      r"^((?:https?:)?\/\/)?((?:www|m)\.)?((?:youtube\.com|youtu.be))(\/(?:[\w\-]+\?v=|embed\/|v\/)?)([\w\-]+)(\S+)?$");
-                  if (clipboard != null && clipboard.text != null && youtubeRegEx.hasMatch(clipboard.text!)) {
-                    _addDownloadController.text = clipboard.text!;
-                  }
-                }
-                showPopoverWB(
-                  context: context,
-                  onConfirm: () {
-                    context.back();
-                    if (_addDownloadController.value.text.isNotEmpty) {
-                      showDownloadPopup(context, videoUrl: _addDownloadController.text);
-                    }
-                  },
-                  hint: "https://youtube.com/watch?v=***********",
-                  title: "Download from video url",
-                  controller: _addDownloadController,
-                );
-              },
-              child: const Icon(LucideIcons.plus),
-            )
-          : null,
-      bottomNavigationBar: Visibility(
-        visible: context.isMobile,
-        child: Container(
-          margin: const EdgeInsets.fromLTRB(4, 0, 4, 4),
-          padding: const EdgeInsets.all(6),
-          decoration: ShapeDecoration(
-            shape: const StadiumBorder(),
-            color: context.getAltBackgroundColor,
-          ),
-          child: SalomonBottomBar(
-            selectedItemColor: context.textTheme.bodyText1!.color,
-            unselectedItemColor: context.getAlt2BackgroundColor,
-            items: [
-              for (var item in navItems.entries)
-                SalomonBottomBarItem(
-                  title: Text(item.key, style: context.textTheme.bodyText1!.copyWith(fontSize: 15)),
-                  icon: Icon(item.value[0], size: 20),
-                  activeIcon: Icon(item.value.length == 2 ? item.value[1] : item.value[0], size: 20),
-                ),
-            ],
-            currentIndex: _currentIndex.value,
-            onTap: (index) => _controller.jumpToPage(index),
-          ),
-        ),
-      ),
     );
   }
 }
