+++ lib/widgets/build_search_button.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutube/screens/screens.dart';
-import 'package:lucide_icons/lucide_icons.dart';
+import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 
 Widget buildSearchButton(BuildContext context) {
   return IconButton(
     onPressed: () => showSearch(context: context, delegate: CustomSearchDelegate()),
-    icon: const Icon(LucideIcons.search, size: 20),
+    icon: const FaIcon(FontAwesomeIcons.search, size: 20),
   );
 }
