+++ lib/widgets/icon_with_bottom_label.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 
 Widget iconWithBottomLabel({
   required IconData icon,
@@ -10,7 +11,7 @@ Widget iconWithBottomLabel({
     child: Column(
       children: [
         IconButton(
-          icon: Icon(icon, size: 28),
+          icon: FaIcon(icon, size: 28),
           onPressed: onPressed ?? () {},
         ),
         const SizedBox(height: 2),
