+++ pubspec.yaml
@@ -1,9 +1,9 @@
 name: flutube
-description: Youtube video downloader made using flutter.
+description: Youtube client made using flutter.
 
 publish_to: none
 
-version: 0.6.0
+version: 1.0.0-dev.0
 
 environment:
   sdk: ">=2.12.0 <3.0.0"
@@ -20,12 +20,12 @@ dependencies:
   flutter:
     sdk: flutter
   flutter_hooks: ^0.18.0
-  flutter_staggered_grid_view: ^0.4.0
+  flutter_staggered_grid_view: ^0.5.0-dev.1
+  font_awesome_flutter: ^10.0.0-beta
   hive: ^2.0.4
   hive_flutter: ^1.1.0
   hooks_riverpod: ^1.0.0-dev.7
   intl: ^0.17.0
-  lucide_icons: ^0.16.12
   open_file: # Upgrade to pub.dev when package > 3.2.1
     git:
       url: https://github.com/miDeb/open_file.git
